cavalini {Bayesrel}	R Documentation

8-Item Questionnaire Data from Cavalini (1992)

Description

A dataset consisting of eight item questionnaire data. It's likert scaled from 0-3. It is data measuring how annoyed people were by malodors

Usage

cavalini

Format

The format is a 8-column datamatrix containing 828 observations

Source

Doctoral Dissertation

References

Cavalini, P. M. (1992). It's an ill wind that brings no good: Studies on odour annoyance and the dispersion of odorant concentrations from industries. Rijksuniversiteit Groningen.

[Package Bayesrel version 0.6.5.5 Index]