p_strel {Bayesrel}	R Documentation

prior and posterior probability of estimate being bigger than threshold

Description

takes a mcmc posterior sample of any of the single test reliability estimates and calculates the prior and posterior probability of the estimate being bigger or smaller than an arbitrary value (priors are stored in the package)

Usage


p_strel(x, estimate, low.bound)

Arguments

x
A strel output object (list)

estimate
A character string indicating what estimate to plot from the strel output object

low.bound
A number for the threshold to be tested against

Examples


p_strel(strel(asrm, "lambda2", n.chains = 1), "lambda2", .80)

[Package Bayesrel version 0.6.5.1 Index]