plot_strel {Bayesrel}	R Documentation

plot function for an single test reliability estimate's posterior sample

Description

gives posterior and prior distribution and pie plots input is the main reliability estimation object and the estimate to be plotted

Usage


plot_strel(
  x,
  estimate,
  blackwhite = FALSE,
  criteria = TRUE,
  cuts = c(0.7, 0.8)
)

Arguments

x
A strel output object (list)

estimate
A character string indicating what estimate to plot from the strel output object

blackwhite
A logical indicating if the plot should be in black and white

criteria
A logical indicating if cutoff criteria should be drawn

cuts
A two element vector indicating what the cutoffs should be

Examples


plot_strel(strel(asrm, "lambda2", n.chains = 2, freq = FALSE), "lambda2")


[Package Bayesrel version 0.6.5.3 Index]